# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RoboKitchen Ordering System",
    "description": "Backend for the RoboKitchen QR ordering app, staff panel, admin panel and robot kitchen.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/staff/kitchen/jobs": {
      "get": {
        "tags": [
          "Staff panel"
        ],
        "summary": "Kitchen Jobs",
        "operationId": "kitchen_jobs_api_v1_staff_kitchen_jobs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/staff/kitchen/jobs/{job_id}/status": {
      "post": {
        "tags": [
          "Staff panel"
        ],
        "summary": "Manual Job Status",
        "description": "Manual override when a staff member finishes an order by hand.",
        "operationId": "manual_job_status_api_v1_staff_kitchen_jobs__job_id__status_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KitchenJobStatusIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/kitchen/jobs": {
      "get": {
        "tags": [
          "Robot kitchen"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_api_v1_kitchen_jobs_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "x-kitchen-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Robot kitchen API key",
              "title": "X-Kitchen-Key"
            },
            "description": "Robot kitchen API key"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kitchen/jobs/claim": {
      "post": {
        "tags": [
          "Robot kitchen"
        ],
        "summary": "Claim Next Job",
        "description": "Atomically take the next queued job. Returns 204 when the queue is empty.",
        "operationId": "claim_next_job_api_v1_kitchen_jobs_claim_post",
        "parameters": [
          {
            "name": "station",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 40,
              "default": "robot-1",
              "title": "Station"
            }
          },
          {
            "name": "x-kitchen-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Robot kitchen API key",
              "title": "X-Kitchen-Key"
            },
            "description": "Robot kitchen API key"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kitchen/jobs/{job_id}": {
      "get": {
        "tags": [
          "Robot kitchen"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_v1_kitchen_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "x-kitchen-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Robot kitchen API key",
              "title": "X-Kitchen-Key"
            },
            "description": "Robot kitchen API key"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/kitchen/jobs/{job_id}/status": {
      "post": {
        "tags": [
          "Robot kitchen"
        ],
        "summary": "Update Job",
        "description": "Report progress: cooking -> done (or failed). Order status follows automatically.",
        "operationId": "update_job_api_v1_kitchen_jobs__job_id__status_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "x-kitchen-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Robot kitchen API key",
              "title": "X-Kitchen-Key"
            },
            "description": "Robot kitchen API key"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KitchenJobStatusIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KitchenJobStatusIn": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "cooking | done | failed"
          },
          "message": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "progress": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "KitchenJobStatusIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}